Load window settings from file

+++ source/utopian/core/renderer/jobs/SunShaftJob.cpp
@@ -43,7 +43,7 @@ namespace Utopian
 	void SunShaftJob::Render(const JobInput& jobInput)
 	{
 		// Move sun
-		mSunAzimuth += Timer::Instance().GetTime() / 10000000 * jobInput.renderingSettings.sunSpeed;
+		mSunAzimuth += (float)Timer::Instance().GetTime() / 10000000 * jobInput.renderingSettings.sunSpeed;
 
 		// Calculate light direction
 		float sunInclination = glm::radians(jobInput.renderingSettings.sunInclination);
